src/ftp_sftp/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery
google-cloud-secret-manager
google-cloud-storage
google-cloud-bigquery[pandas]==2.34.3
